Rename `menu` package into `menuDrawer`
