+++ app/src/main/java/com/infomaniak/mail/ui/main/menuDrawer/MenuDrawerFragment.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.ui.main.menu
+package com.infomaniak.mail.ui.main.menuDrawer
 
 import android.annotation.SuppressLint
 import android.content.Intent
