+++ app/src/main/res/layout/bottom_sheet_restore_emails.xml
@@ -19,7 +19,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.menu.RestoreEmailsBottomSheetDialog">
+    tools:context=".ui.main.menuDrawer.RestoreEmailsBottomSheetDialog">
 
     <TextView
         android:id="@+id/title"
