+++ app/src/main/res/layout/fragment_mailbox_settings.xml
@@ -15,7 +15,7 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program.  If not, see <http://www.gnu.org/licenses/>.
   -->
-<com.infomaniak.mail.ui.main.menu.SimpleSettingView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.infomaniak.mail.ui.main.menuDrawer.SimpleSettingView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -170,4 +170,4 @@
 
         </LinearLayout>
     </androidx.core.widget.NestedScrollView>
-</com.infomaniak.mail.ui.main.menu.SimpleSettingView>
+</com.infomaniak.mail.ui.main.menuDrawer.SimpleSettingView>
